# docs/swagger.yaml
definitions:
  gin.H:
    additionalProperties: {}
    type: object
...
  github_com_EvolutionAPI_evolution-go_pkg_instance_model.AdvancedSettings:
    properties:
      alwaysOnline:
        type: boolean
      ignoreGroups:
        type: boolean
      ignoreStatus:
        type: boolean
      msgRejectCall:
        type: string
      readMessages:
        type: boolean
      rejectCall:
        type: boolean
    type: object
  github_com_EvolutionAPI_evolution-go_pkg_instance_service.ConnectStruct:
    properties:
      immediate:
        type: boolean
      natsEnable:
        type: string
      phone:
        type: string
      rabbitmqEnable:
        type: string
      subscribe:
        items:
          type: string
        type: array
      webhookUrl:
        type: string
      websocketEnable:
        type: string
    type: object
  github_com_EvolutionAPI_evolution-go_pkg_instance_service.CreateStruct:
    properties:
      advancedSettings:
        $ref: '#/definitions/github_com_EvolutionAPI_evolution-go_pkg_instance_model.AdvancedSettings'
      instanceId:
        type: string
      name:
        type: string
      os_name:
        type: string
      proxy:
        $ref: '#/definitions/github_com_EvolutionAPI_evolution-go_pkg_instance_service.ProxyConfig'
      token:
        type: string
    type: object
  github_com_EvolutionAPI_evolution-go_pkg_instance_service.ForceReconnectStruct:
    properties:
      number:
        type: string
    type: object
  github_com_EvolutionAPI_evolution-go_pkg_instance_service.PairStruct:
    properties:
      phone:
        type: string
      subscribe:
        items:
          type: string
        type: array
    type: object
  github_com_EvolutionAPI_evolution-go_pkg_instance_service.ProxyConfig:
    properties:
      host:
        type: string
      password:
        type: string
      port:
        type: string
      username:
        type: string
    type: object
  github_com_EvolutionAPI_evolution-go_pkg_instance_service.SetProxyStruct:
    properties:
      host:
        type: string
      password:
        type: string
      port:
        type: string
      username:
        type: string
    required:
    - host
    - port
    type: object
...
info:
  contact: {}
  description: Evolution GO - whatsmeow
  title: Evolution GO
  version: "1.0"
paths:
...
  /instance/{instanceId}/advanced-settings:
    get:
      description: Get advanced settings for a specific instance
      parameters:
      - description: Instance ID
        in: path
        name: instanceId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Advanced settings retrieved successfully
          schema:
            $ref: '#/definitions/github_com_EvolutionAPI_evolution-go_pkg_instance_model.AdvancedSettings'
        "400":
          description: Invalid instance ID
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: Instance not found
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Get advanced settings
      tags:
      - Instance
    put:
      consumes:
      - application/json
      description: Update advanced settings for a specific instance
      parameters:
      - description: Instance ID
        in: path
        name: instanceId
        required: true
        type: string
      - description: Advanced settings data
        in: body
        name: settings
        required: true
        schema:
          $ref: '#/definitions/github_com_EvolutionAPI_evolution-go_pkg_instance_model.AdvancedSettings'
      produces:
      - application/json
      responses:
        "200":
          description: Advanced settings updated successfully
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Invalid request data
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: Instance not found
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Update advanced settings
      tags:
      - Instance
  /instance/all:
    get:
      consumes:
      - application/json
      description: Get all instances
      produces:
      - application/json
      responses:
        "200":
          description: All instances
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Get all instances
      tags:
      - Instance
  /instance/connect:
    post:
      consumes:
      - application/json
      description: Connect to instance with the provided data
      parameters:
      - description: Instance data
        in: body
        name: instance
        required: true
        schema:
          $ref: '#/definitions/github_com_EvolutionAPI_evolution-go_pkg_instance_service.ConnectStruct'
      produces:
      - application/json
      responses:
        "200":
          description: Instance connected successfully
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Error on validation
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Connect to instance
      tags:
      - Instance
  /instance/create:
    post:
      consumes:
      - application/json
      description: Creates a new instance with the provided data including optional
        os_name and advanced settings
      parameters:
      - description: Instance data with optional os_name and advanced settings
        in: body
        name: instance
        required: true
        schema:
          $ref: '#/definitions/github_com_EvolutionAPI_evolution-go_pkg_instance_service.CreateStruct'
      produces:
      - application/json
      responses:
        "200":
          description: Instance created successfully
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Error on validation
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Create a new instance
      tags:
      - Instance
  /instance/delete/{instanceId}:
    delete:
      consumes:
      - application/json
      description: Delete instance
      parameters:
      - description: Instance Id
        in: path
        name: instanceId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Instance deleted successfully
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Error on validation
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Delete instance
      tags:
      - Instance
  /instance/disconnect:
    post:
      consumes:
      - application/json
      description: Disconnect from instance
      produces:
      - application/json
      responses:
        "200":
          description: Instance disconnected successfully
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Disconnect from instance
      tags:
      - Instance
  /instance/forcereconnect/{instanceId}:
    post:
      consumes:
      - application/json
      description: Force reconnect
      parameters:
      - description: Instance Id
        in: path
        name: instanceId
        required: true
        type: string
      - description: Instance data
        in: body
        name: instance
        required: true
        schema:
          $ref: '#/definitions/github_com_EvolutionAPI_evolution-go_pkg_instance_service.ForceReconnectStruct'
      produces:
      - application/json
      responses:
        "200":
          description: Instance force reconnected successfully
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Error on validation
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Force reconnect
      tags:
      - Instance
  /instance/get/{instanceId}:
    get:
      consumes:
      - application/json
      description: Get instance
      parameters:
      - description: Instance Id
        in: path
        name: instanceId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Instance
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Error on validation
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Get instance
      tags:
      - Instance
  /instance/logout:
    delete:
      consumes:
      - application/json
      description: Logout from instance
      produces:
      - application/json
      responses:
        "200":
          description: Instance logged out successfully
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Logout from instance
      tags:
      - Instance
  /instance/pair:
    post:
      consumes:
      - application/json
      description: Request pairing code
      parameters:
      - description: Instance data
        in: body
        name: instance
        required: true
        schema:
          $ref: '#/definitions/github_com_EvolutionAPI_evolution-go_pkg_instance_service.PairStruct'
      produces:
      - application/json
      responses:
        "200":
          description: Pairing code
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Error on validation
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Request pairing code
      tags:
      - Instance
  /instance/proxy/{instanceId}:
    delete:
      consumes:
      - application/json
      description: Delete proxy
      parameters:
      - description: Instance id
        in: path
        name: instanceId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Proxy deleted successfully
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Error on validation
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Delete proxy
      tags:
      - Instance
    post:
      consumes:
      - application/json
      description: Set proxy configuration for an instance
      parameters:
      - description: Instance id
        in: path
        name: instanceId
        required: true
        type: string
      - description: Proxy configuration
        in: body
        name: proxy
        required: true
        schema:
          $ref: '#/definitions/github_com_EvolutionAPI_evolution-go_pkg_instance_service.SetProxyStruct'
      produces:
      - application/json
      responses:
        "200":
          description: Proxy set successfully
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Error on validation
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Set proxy configuration
      tags:
      - Instance
  /instance/qr:
    get:
      consumes:
      - application/json
      description: Get instance QR code
      produces:
      - application/json
      responses:
        "200":
          description: Instance QR code
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Get instance QR code
      tags:
      - Instance
  /instance/reconnect:
    post:
      consumes:
      - application/json
      description: Reconnect to instance
      produces:
      - application/json
      responses:
        "200":
          description: Instance reconnected successfully
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Reconnect to instance
      tags:
      - Instance
  /instance/status:
    get:
      consumes:
      - application/json
      description: Get instance status
      produces:
      - application/json
      responses:
        "200":
          description: Instance status
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Get instance status
      tags:
      - Instance
...
swagger: "2.0"